+++ src/main.ts
@@ -1,9 +1,11 @@
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
-import Vant, { Toast } from 'vant'
+import Vant from 'vant'
 import 'vant/lib/index.css'
 import App from './App.vue'
 import router from './router'
+import headicon from '@/plugins/headicon'
+import content from '@/plugins/content'
 import './styles/global.scss'
 
 const app = createApp(App)
@@ -11,9 +13,7 @@ const app = createApp(App)
 app.use(createPinia())
 app.use(router)
 app.use(Vant)
-
-app.config.globalProperties.$toast = (message: string, type: string = 'fail') => {
-  Toast(type === 'success' ? { message, type: 'success' } : { message, type })
-}
+app.use(headicon)
+app.use(content)
 
 app.mount('#app')